+++ benchmark/index.ts
@@ -12,7 +12,7 @@ interface TransformedEntry {
 const COUNT = 1000000;
 const REQUEST = request;
 REQUEST.headers = {
-    authentication: 'HMAC 1573504737300:76251c6323fbf6355f23816a4c2e12edfd10672517104763ab1b10f078277f86'
+    authorization: 'HMAC 1573504737300:76251c6323fbf6355f23816a4c2e12edfd10672517104763ab1b10f078277f86'
 };
 REQUEST.method = 'POST';
 REQUEST.originalUrl = '/api/order';